Added Pronounfluid and Pronounflux
Pronounfluid is the act of switching or transitioning between two or more pronoun sets over time. One's pronouns may change over the course of hours, days, weeks or months, and the gender of the person does not matter, and does not have to change with the pronouns.

Pronounflux is when one's pronoun preference fluctuates in intensity. An example of being pronounflux is where someone's pronouns are she/them/xem, and one day they prefer "she/her" over any other pronoun in their list, the next day they prefer "they/them". Another example is when a person has a batch of pronouns they use, but how much they like them might fluctuate.