Added Catgender, Catnipgender, and Dragongender
Catgender is a xenogender in which someone feels an extremely strong connection to cats or other felines, either strongly identifying with them or simply wanting to incorporate them into their gender to better understand their identity. It is used to describe a gender that feels feline.

Catnipgender is a subgender in which an individual is basically catgender, but also experiences other connections to genders which may be more silly, hyperactive, funny, crazy, and/or wacky at the same time, or could just be attached to of the concepts in overall.

Dragongender is a xenogender gender identity on which is reflected by dragons.