Added 6 gender flags
Gxrl is a gender where a female and an agender identity are experienced simultaneously or with variation.

Bxy is a gender where a male and an agender identity are experienced simultaneously or with variation.

Non-binary woman, or non-binary girl, refers to someone who identifies both as a woman and as non-binary. 

Non-binary man, or non-binary boy, refers to someone who identifies both as a man and as non-binary.