Added 2 aroace flags
Nebularomantic is a romantic orientation on the aromantic spectrum. It is described when one is unable to or has a hard time distinguishing romantic attraction from platonic attraction due to being neurodivergent.

Nebulasexual is a sexuality under the quoisexual umbrella where one cannot tell if they experience sexual attraction or not due to neurodivergency. One who is nebulasexual might want sex or a sexual relationship, but they do not know if they experience attraction.